pyproject.toml
[project]
name = "pubmed-fetcher"
version = "0.1.0"
description = "Fetch PubMed papers with biotech affiliations"
authors = [
    {name = "Bharani ",email = "bharanichandhran@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "typer (>=0.16.0,<0.17.0)",
    "biopython (>=1.85,<2.0)"
]
[tool.poetry.scripts]
get-papers-list = "cli:main"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
